Extract all things BitmapRef into a dedicated module
